fix(participant): use precise monetary amounts

+++ src/main/kotlin/com/stockcomp/participant/InvestmentDto.kt
@@ -1,13 +1,14 @@
 package com.stockcomp.participant
 
 import com.stockcomp.participant.internal.investment.Investment
+import java.math.BigDecimal
 
 data class InvestmentDto(
     val symbol: String,
     val amount: Int,
-    val averageUnitCost: Double,
-    val totalProfit: Double,
-    val totalValue: Double,
+    val averageUnitCost: BigDecimal,
+    val totalProfit: BigDecimal,
+    val totalValue: BigDecimal,
 )
 
 fun mapToInvestmentDto(source: Investment) =
