+++ src/main/kotlin/com/stockcomp/participant/InvestmentOrderDto.kt
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.Future
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Positive
 import jakarta.validation.constraints.Size
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class PlaceInvestmentOrderRequest(
@@ -22,7 +23,7 @@ data class PlaceInvestmentOrderRequest(
     @field:Future
     val expirationTime: LocalDateTime,
     @field:Positive
-    val acceptedPrice: Double,
+    val acceptedPrice: BigDecimal,
     val transactionType: TransactionType,
 )
 
@@ -31,7 +32,7 @@ data class InvestmentOrderDto(
     val symbol: String,
     val totalAmount: Int,
     val remainingAmount: Int,
-    val acceptedPrice: Double,
+    val acceptedPrice: BigDecimal,
     val currency: String,
     val expirationTime: LocalDateTime,
     val transactionType: TransactionType,
